backend/pyproject.toml
[project]
name = "chatkit-backend"
version = "0.1.0"
description = "FastAPI backend service for ChatKit boilerplate"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "httpx>=0.28,<0.29",
    "uvicorn[standard]>=0.36,<0.37",
    "openai>=1.40",
    "openai-chatkit>=1.0.2,<2",
    "boto3>=1.34,<2",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.4,<0.7",
    "mypy>=1.8,<2",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["py.typed"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]
